Update to 1.21, port to NeoForge
