+++ common/src/main/resources/roughlyenoughresources.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "package": "uk.me.desert_island.rer.mixin",
-  "compatibilityLevel": "JAVA_17",
+  "compatibilityLevel": "JAVA_21",
   "mixins": ["IdentifierHooks", "MixinChunkGenerator", "MixinMinecraftServer", "MixinPlayerManager", "MixinServerWorld"],
   "client": [
     "MixinClientConnection"
