+++ fabric/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id "com.github.johnrengelman.shadow" version "7.1.2"
+    id "com.github.johnrengelman.shadow" version "8.1.1"
 }
 
 architectury {
@@ -22,10 +22,9 @@ dependencies {
     // PSA: Some older mods, compiled on Loom 0.2.1, might have outdated Maven POMs.
     // You may need to force-disable transitiveness on them.
 
-    // TODO: this should actually be modCompileOnly "me.shedaniel:RoughlyEnoughItems-api:..." but RER depends on the internal EntryWidget
     modCompileOnly "me.shedaniel:RoughlyEnoughItems-api-fabric:${project.rei_version}"
     modImplementation "me.shedaniel:RoughlyEnoughItems-fabric:${project.rei_version}"
-    modImplementation "dev.architectury:architectury-fabric:3.2.51"
+    modImplementation "dev.architectury:architectury-fabric:${project.architectury_version}"
     modImplementation("me.shedaniel.cloth:cloth-config-fabric:$cloth_config_version") {
         exclude "module": "fabric-api"
     }
@@ -46,7 +45,7 @@ shadowJar {
     exclude "architectury.common.json"
 
     configurations = [project.configurations.shadowCommon]
-    classifier "dev-shadow"
+    archiveClassifier.set "dev-shadow"
 }
 
 
@@ -54,19 +53,11 @@ java {
     withSourcesJar()
 }
 
-jar {
-    from "LICENSE"
-}
-
 remapJar {
     injectAccessWidener = true
     input.set shadowJar.archiveFile
     dependsOn shadowJar
-    classifier null
-}
-
-jar {
-    classifier "dev"
+    archiveClassifier.set ""
 }
 
 sourcesJar {
