--- forge/gradle.properties
@@ -1,1 +1,0 @@
-loom.platform=forge
