--- forge/src/main/resources/pack.mcmeta
@@ -1,6 +1,0 @@
-{
-  "pack": {
-    "description": "Roughly Enough Resources",
-    "pack_format": 15
-  }
-}
