+++ gradle.properties
@@ -2,26 +2,16 @@
 org.gradle.jvmargs=-Xmx3G
 # Fabric Properties
 # check these on https://modmuss50.me/fabric.html
-minecraft_version=1.20.1
-fabric_loader_version=0.14.21
-forge_version=1.20.1-47.0.19
+minecraft_version=1.21.1
+enabled_platforms=fabric,neoforge
+fabric_loader_version=0.16.7
+neoforge_version=21.1.62
 # Mod Properties
-mod_version=2.9.0
+mod_version=3.0.0
 maven_group=net.fabricmc
 archives_base_name=rer
 # Dependencies
-fabric_version=0.83.1+1.20.1
-cloth_config_version=11.0.99
+fabric_version=0.107.0+1.21.1
+cloth_config_version=15.0.140
 architectury_version=9.0.8
-#arcanemagic_version=1.0.0-beta4
-# copied from arcane magic
-# Required Dependencies
-#silk_version=1.2.4-43
-#multiblock_api_version=1.0.1:1
-#cutscene_api_version=1.0.0:beta5
-#crochet_version=1.0.4
-# Optional Dependencies
-rei_version=12.0.625
-#modmenu_version=1.10.1+build.30
-#informedload_version=1.1.1
-#autoconfig_version=1.1.2+mc1.14.1
+rei_version=16.0.754
