+++ neoforge/src/main/java/me/desert_island/rer/neoforge/PluginEntryNeoforge.java
@@ -1,7 +1,8 @@
-package uk.me.desert_island.rer.forge;
+package me.desert_island.rer.neoforge;
 
 import me.shedaniel.rei.forge.REIPluginCommon;
 import uk.me.desert_island.rer.rei_stuff.PluginEntry;
 
 @REIPluginCommon
-public class PluginEntryForge extends PluginEntry {}
+public class PluginEntryNeoforge extends PluginEntry {
+}
